pyproject.toml
[project]
name = "image-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.1",
    "google-genai>=1.66.0",
    "pydantic>=2.0",
    "python-dotenv>=1.2.2",
    "webptools>=0.0.9",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.21.0",
]
